+++ ICD.Connect.Telemetry.Services/Properties/AssemblyInfo.cs
@@ -4,4 +4,4 @@
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Telemetry.Services")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2021")]
-[assembly: AssemblyVersion("8.2.2.0")]
+[assembly: AssemblyVersion("8.3.0.0")]