+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
- - Changed Program/System uptime to start time
 
+## [8.3.0] - 2021-02-04
 ### Added
  - Added debounce support to event telemetry items
+
+### Changed
+ - Changed Program/System uptime to start time
 
 ## [8.2.2] - 2021-01-25
 ### Added